Using int and IntByReference for ULONG, also added ULONG.

+++ contrib/platform/src/com/sun/jna/platform/win32/Secur32.java
@@ -95,7 +95,7 @@ public interface Secur32 extends StdCallLibrary {
      *  If the function fails, the function returns one of the SEC_E_ error codes.
      */
     public int AcquireCredentialsHandle(String pszPrincipal, String pszPackage,
-                                        long fCredentialUse, LUID pvLogonID,
+                                        int fCredentialUse, LUID pvLogonID,
                                         Pointer pAuthData, Pointer pGetKeyFn, // TODO: SEC_GET_KEY_FN
                                         Pointer pvGetKeyArgument, CredHandle phCredential, 
                                         TimeStamp ptsExpiry);
@@ -162,8 +162,8 @@ public interface Secur32 extends StdCallLibrary {
      *  If the function fails, the function returns one of the SEC_E_ error codes.
      */
     public int InitializeSecurityContext(CredHandle phCredential, CtxtHandle phContext,
-                                         String pszTargetName, long fContextReq, long Reserved1,
-                                         long TargetDataRep, SecBufferDesc pInput, long Reserved2,
+                                         String pszTargetName, int fContextReq, int Reserved1,
+                                         int TargetDataRep, SecBufferDesc pInput, int Reserved2,
                                          CtxtHandle phNewContext, SecBufferDesc pOutput, IntByReference pfContextAttr,
                                          TimeStamp ptsExpiry);
 	
@@ -236,7 +236,7 @@ public interface Secur32 extends StdCallLibrary {
      *  This function returns one of SEC_* values.
      */
     public int AcceptSecurityContext(CredHandle phCredential, CtxtHandle phContext,
-                                     SecBufferDesc pInput, long fContextReq, long TargetDataRep,
+                                     SecBufferDesc pInput, int fContextReq, int TargetDataRep,
                                      CtxtHandle phNewContext, SecBufferDesc pOutput, IntByReference pfContextAttr,
                                      TimeStamp ptsTimeStamp);
 
@@ -244,7 +244,7 @@ public interface Secur32 extends StdCallLibrary {
      * The EnumerateSecurityPackages function returns an array of SecPkgInfo structures that 
      * describe the security packages available to the client.
      * @param pcPackages
-     *  A pointer to a ULONG variable that receives the number of packages returned.
+     *  A pointer to a int variable that receives the number of packages returned.
      * @param ppPackageInfo
      *  A pointer to a variable that receives a pointer to an array of SecPkgInfo structures. 
      *  Each structure contains information from the security support provider (SSP) that
